ceph file: tidy up file/mds operator code
Separate `file` operator from `mds` tasks logically for easier code
navigation.
